requirements.txt
mkdocs>=1.6.0,<2.0.0
mkdocs-material>=9.5.0,<10.0.0
mkdocs-macros-plugin>=1.0.5,<2.0.0
PyYAML>=6.0.1
